requirements.txt
pyTelegramBotAPI
gspread
google-auth
python-dotenv
flask
fpdf2
